Split object file generation out from LLVMGen
